+++ src/source-translation.js
@@ -1,5 +1,6 @@
 const crypto = require('crypto');
 const path = require('path');
+const { compare, getValueByPointer } = require('fast-json-patch');
 const { fileClass } = require('./utils');
 
 class SourceTranslation {
@@ -27,8 +28,56 @@ class SourceTranslation {
     await this.hash.read();
   }
 
-  hashIt() {
-    this.newHashData = this._hashThisObject(this.file.data);
+  buildChanges() {
+    let newHashData = this._hashThisObject(this.file.data);
+    let operations = compare(this.hash.data, newHashData);
+    let changes = [];
+
+    operations.forEach(operation => {
+      if (operation.op === 'remove') {
+        changes.push({
+          op: 'remove',
+          path: operation.path,
+        });
+      } else if (operation.op === 'add' || operation.op === 'replace') {
+        this._addTranslationChanges(operation, changes);
+      }
+    });
+
+    return changes;
+  }
+
+  _addTranslationChangeForPath(path, changes) {
+    changes.push({
+      op: 'translate',
+      path: path,
+      sourceTranslation: getValueByPointer(this.file.data, path),
+    });
+  }
+
+  _addTranslationChanges(operation, changes) {
+    if (typeof operation.value === 'object') {
+      this._expandTranslationChanges(operation.value, operation.path, changes);
+    } else if (typeof operation.value === 'string') {
+      this._addTranslationChangeForPath(operation.path, changes);
+    }
+  }
+
+  _expandTranslationChanges(object, path, changes) {
+    for (let key in object) {
+      if (Object.prototype.hasOwnProperty.call(object, key)) {
+        let value = object[key];
+        let deeperPath = `${path}/${key}`;
+
+        if (typeof value === 'object') {
+          this._expandTranslationChanges(value, deeperPath, changes);
+        } else if (typeof value === 'string') {
+          this._addTranslationChangeForPath(deeperPath, changes);
+        } else {
+          throw new TypeError('unable to expand changes for unknown value type');
+        }
+      }
+    }
   }
 
   _hashThisObject(obj) {
@@ -39,10 +88,12 @@ class SourceTranslation {
         let value = obj[key];
 
         if (Array.isArray(value)) {
+          // I don't think this is actually a real case. Consider removing/erroring?
           newObj[key] = value.map(item => this._hashThisObject(item));
         } else if (typeof value === 'object') {
           newObj[key] = this._hashThisObject(value);
         } else if (typeof value === 'string') {
+          // here's the real deal
           newObj[key] = crypto.createHash('sha1').update(value).digest('base64');
         } else {
           throw new TypeError('unable to hash unknown value type');