+++ src/json-file.js
@@ -11,9 +11,11 @@ class JsonFile {
     try {
       this.file = await fs.promises.readFile(this.filepath);
       this.data = JSON.parse(this.file);
+      this.exists = true;
     } catch (error) {
       if (error.code === 'ENOENT' && Boolean(this.optional)) {
         this.data = {};
+        this.exists = false;
       } else {
         throw error;
       }