+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Loading YAML and JSON source translations, their hash file, and calculating the new hash data
 - Comparing the old and new hash data via JSON Patch, and transforming that into a set of translation change templates
 - Compiling a full set of changes to make linked to each specific destination translation file
+- Logging messages informing the user about what we loaded, what we couldn't load, how many changes we've detected, etc
 
 ## [0.2.0] - 2020-04-26